third_party/WebKit/LayoutTests/http/tests/security/cross-frame-access-callback-explicit-domain-DENY.html

<html>
<body>
<p>Test that a child frame can't define a function and the use it to access parent properties after document.domain write blocks the access.
<script>
if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpChildFrames();
}
</script>
<iframe src="resources/cross-frame-iframe-callback-explicit-domain-DENY.html" style=""></iframe>
<pre id="console"></pre>
</body>
</html>
