third_party/blink/web_tests/http/tests/security/cross-frame-access-callback-explicit-domain-DENY-expected.txt

Test that a child frame can't define a function and the use it to access parent properties after document.domain write blocks the access. 

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS: canGet('parentWindow.location.href') should be 'false' and is.
PASS: accessThrowsException('parentWindow.location.href') should be 'true' and is.
